Revert "Merge MethodSignatureProcessor into JarTransformerChain (#30)"
This reverts commit 61aa594d345b129e0412a263996ff3ed129a8a5b.
